corrected errors, made more user friendly
-added value 0 for column rx_prov_title in ProviderPreference table.

- had if inactive_staff_use: for doctor accounts
& had if inactive_doctor_use: for staff accounts
Correct this so they use the proper inactive booleans.


- clinic array
Changed it so you can leave it blank if it is a single office clinic and addprov will properly create the 'Doctors' group that we are used to having.


- messenger group

Somewhere along the way I deleted the line that actually updated oscarcommlocations for messenger groups. I have put this line back in.